requirements.txt
numpy>=1.23,<2.0
opencv-python>=4.8,<5.0
scipy>=1.9,<2.0
scikit-image>=0.20,<0.23
pillow>=11.1,<13.0
pillow-heif>=1.2,<2.0
streamlit>=1.31,<2.0

pyyaml>=6.0
PySide6>=6.6,<7.0

matplotlib>=3.7,<4.0
tqdm>=4.65

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
